# Adding compatibility  with prometheus 'stringlabels' build flag