--- a/.github/workflows/test.yaml
+++ b/.github/workflows/test.yaml
@@ -68,3 +68,21 @@ jobs:
 
       - name: Run unit tests
         run: make test
+
+  test-tag-stringlabels:
+    runs-on: ubuntu-latest
+    name: Run tests --tags=stringlabels
+    env:
+      GOBIN: /tmp/.bin
+    steps:
+      - name: Check out code into the Go module directory.
+        uses: actions/checkout@v3
+
+      - name: Install Go.
+        uses: actions/setup-go@v3
+        with:
+          go-version-file: .go-version
+          cache: true
+
+      - name: Run unit tests
+        run: make test-stringlabels