--- a/Makefile
+++ b/Makefile
@@ -37,6 +37,15 @@ test:
 	@rm -rf $(GOCACHE)
 	@go test -v -race -timeout=10m $(shell go list ./...);
 
+
+.PHONY: test-stringlabels
+test-stringlabels: ## Runs all Go unit tests with stringlabels flag.
+export GOCACHE=/tmp/cache
+test-stringlabels:
+	@echo ">> stringlabels: running unit tests (without cache)"
+	@rm -rf $(GOCACHE)
+	@go test -v -race --tags=stringlabels -timeout=10m $(shell go list ./...);
+
 .PHONY: deps
 deps: ## Ensures fresh go.mod and go.sum.
 	@go mod tidy