--- a/engine/engine_test.go
+++ b/engine/engine_test.go
@@ -16,6 +16,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/google/go-cmp/cmp"
 	promparser "github.com/prometheus/prometheus/promql/parser"
 
 	"github.com/google/go-cmp/cmp/cmpopts"
@@ -1736,7 +1737,7 @@ func TestQueriesAgainstOldEngine(t *testing.T) {
 								testutil.Ok(t, newResult.Err)
 								if hasNaNs(oldResult) {
 									t.Log("Applying comparison with NaN equality.")
-									testutil.WithGoCmp(cmpopts.EquateNaNs()).Equals(t, oldResult, newResult)
+									equalsWithNaNs(t, oldResult, newResult)
 								} else {
 									emptyLabelsToNil(oldResult)
 									emptyLabelsToNil(newResult)
@@ -1748,6 +1749,14 @@ func TestQueriesAgainstOldEngine(t *testing.T) {
 				}
 			})
 		}
+	}
+}
+
+func equalsWithNaNs(t *testing.T, oldResult, newResult interface{}) {
+	if reflect.TypeOf(labels.Labels{}).Kind() == reflect.Struct {
+		testutil.WithGoCmp(cmpopts.EquateNaNs(), cmp.AllowUnexported(labels.Labels{})).Equals(t, oldResult, newResult)
+	} else {
+		testutil.WithGoCmp(cmpopts.EquateNaNs()).Equals(t, oldResult, newResult)
 	}
 }
 
@@ -3309,7 +3318,7 @@ func TestInstantQuery(t *testing.T) {
 
 								if hasNaNs(oldResult) {
 									t.Log("Applying comparison with NaN equality.")
-									testutil.WithGoCmp(cmpopts.EquateNaNs()).Equals(t, oldResult, newResult)
+									equalsWithNaNs(t, oldResult, newResult)
 								} else if oldResult.Err != nil {
 									testutil.Equals(t, oldResult.Err.Error(), newResult.Err.Error())
 								} else {
@@ -4163,7 +4172,7 @@ func testNativeHistograms(t *testing.T, cases []histogramTestCase, opts promql.E
 						sortByLabels(newResult)
 						if hasNaNs(promResult) {
 							t.Log("Applying comparison with NaN equality.")
-							testutil.WithGoCmp(cmpopts.EquateNaNs()).Equals(t, promVector, newVector)
+							equalsWithNaNs(t, promVector, newVector)
 						} else {
 							testutil.Equals(t, promVector, newVector)
 						}
@@ -4192,7 +4201,7 @@ func testNativeHistograms(t *testing.T, cases []histogramTestCase, opts promql.E
 						testutil.Equals(t, len(expected), len(actual))
 						if hasNaNs(res) {
 							t.Log("Applying comparison with NaN equality.")
-							testutil.WithGoCmp(cmpopts.EquateNaNs()).Equals(t, expected, actual)
+							equalsWithNaNs(t, expected, actual)
 						} else {
 							testutil.Equals(t, expected, actual)
 						}