--- a/engine/enginefuzz_test.go
+++ b/engine/enginefuzz_test.go
@@ -150,8 +150,8 @@ func FuzzEnginePromQLSmithInstantQuery(f *testing.F) {
 			return
 		}
 		load := fmt.Sprintf(`load 30s
-			http_requests_total{pod="nginx-1", route="/"} %.2f+%.2fx4
-			http_requests_total{pod="nginx-2", route="/"} %2.f+%.2fx4`, initialVal1, inc1, initialVal2, inc2)
+			http_requests_total{pod="nginx-1", route="/"} %.2f+%.2fx40
+			http_requests_total{pod="nginx-2", route="/"} %2.f+%.2fx40`, initialVal1, inc1, initialVal2, inc2)
 
 		opts := promql.EngineOpts{
 			Timeout:              1 * time.Hour,
@@ -179,6 +179,7 @@ func FuzzEnginePromQLSmithInstantQuery(f *testing.F) {
 		psOpts := []promqlsmith.Option{
 			promqlsmith.WithEnableOffset(true),
 			promqlsmith.WithEnableAtModifier(true),
+			promqlsmith.WithAtModifierMaxTimestamp(180 * 1000),
 		}
 		ps := promqlsmith.New(rnd, seriesSet, psOpts...)
 
@@ -294,6 +295,7 @@ func FuzzDistributedEnginePromQLSmithRangeQuery(f *testing.F) {
 		psOpts := []promqlsmith.Option{
 			promqlsmith.WithEnableOffset(true),
 			promqlsmith.WithEnableAtModifier(true),
+			promqlsmith.WithAtModifierMaxTimestamp(180 * 1000),
 			promqlsmith.WithEnabledAggrs([]parser.ItemType{parser.SUM, parser.MIN, parser.MAX, parser.GROUP, parser.COUNT, parser.BOTTOMK, parser.TOPK}),
 		}
 		ps := promqlsmith.New(rnd, seriesSet, psOpts...)
@@ -398,6 +400,7 @@ func FuzzDistributedEnginePromQLSmithInstantQuery(f *testing.F) {
 		psOpts := []promqlsmith.Option{
 			promqlsmith.WithEnableOffset(true),
 			promqlsmith.WithEnableAtModifier(true),
+			promqlsmith.WithAtModifierMaxTimestamp(180 * 1000),
 			promqlsmith.WithEnabledAggrs([]parser.ItemType{parser.SUM, parser.MIN, parser.MAX, parser.GROUP, parser.COUNT, parser.BOTTOMK, parser.TOPK}),
 		}
 		ps := promqlsmith.New(rnd, seriesSet, psOpts...)