--- a/execution/binary/scalar.go
+++ b/execution/binary/scalar.go
@@ -163,7 +163,7 @@ func (o *scalarOperator) loadSeries(ctx context.Context) error {
 	}
 	series := make([]labels.Labels, len(vectorSeries))
 	for i := range vectorSeries {
-		if vectorSeries[i] != nil {
+		if !vectorSeries[i].IsEmpty() {
 			lbls := vectorSeries[i]
 			if shouldDropMetricName(o.opType, o.returnBool) {
 				lbls, _ = function.DropMetricName(lbls.Copy())