--- a/execution/function/operator.go
+++ b/execution/function/operator.go
@@ -21,11 +21,6 @@ import (
 	"github.com/thanos-community/promql-engine/query"
 )
 
-const (
-	// ExpectedLabelsSize Expected number of labels on a timeseries used as initial value when creating a new labels.ScratchBuilder.
-	ExpectedLabelsSize = 10
-)
-
 // functionOperator returns []model.StepVector after processing input with desired function.
 type functionOperator struct {
 	funcExpr *parser.Call
@@ -209,7 +204,7 @@ func (o *functionOperator) Next(ctx context.Context) ([]model.StepVector, error)
 		o.nextOps[i].GetPool().PutVectors(scalarVectors)
 		scalarIndex++
 	}
-	lblsBuilder := labels.NewScratchBuilder(ExpectedLabelsSize)
+	lblsBuilder := labels.ScratchBuilder{}
 	for batchIndex, vector := range vectors {
 		// scalar() depends on number of samples per vector and returns NaN if len(samples) != 1.
 		// So need to handle this separately here, instead of going via call which is per point.
@@ -276,16 +271,6 @@ func (o *functionOperator) Next(ctx context.Context) ([]model.StepVector, error)
 	return vectors, nil
 }
 
-func (o *functionOperator) newFunctionArgs(vector model.StepVector, batchIndex int, b labels.ScratchBuilder) FunctionArgs {
-	return FunctionArgs{
-		Labels:        o.series[0],
-		Samples:       o.sampleBuf,
-		StepTime:      vector.T,
-		ScalarPoints:  o.scalarPoints[batchIndex],
-		LabelsBuilder: b,
-	}
-}
-
 func (o *functionOperator) loadSeries(ctx context.Context) error {
 	var err error
 	o.once.Do(func() {
@@ -322,7 +307,7 @@ func (o *functionOperator) loadSeries(ctx context.Context) error {
 				labelJoinSrcLabels = append(labelJoinSrcLabels, o.funcExpr.Args[j].(*parser.StringLiteral).Val)
 			}
 		}
-		b := labels.NewScratchBuilder(ExpectedLabelsSize)
+		b := labels.ScratchBuilder{}
 		for i, s := range series {
 			lbls := s
 			switch o.funcExpr.Func.Name {