--- a/execution/scan/matrix_selector.go
+++ b/execution/scan/matrix_selector.go
@@ -205,6 +205,7 @@ func (o *matrixSelector) loadSeries(ctx context.Context) error {
 
 		o.scanners = make([]matrixScanner, len(series))
 		o.series = make([]labels.Labels, len(series))
+		b := labels.NewScratchBuilder(function.ExpectedLabelsSize)
 		for i, s := range series {
 			lbls := s.Labels()
 			if o.funcExpr.Func.Name != "last_over_time" {
@@ -213,7 +214,7 @@ func (o *matrixSelector) loadSeries(ctx context.Context) error {
 				// we have to copy it here.
 				// TODO(GiedriusS): could we identify somehow whether labels.Labels
 				// is reused between Select() calls?
-				lbls, _ = function.DropMetricName(lbls.Copy())
+				lbls, _ = function.DropMetricName(lbls.Copy(), b)
 			}
 
 			// If we are dealing with an extended range function we need to search further in the past for valid series.