Making the engine compatible with the new `stringlabels` prometheus build flag.

Please note that i removed the slice.sort [here](https://github.com/thanos-community/promql-engine/compare/main...alanprot:stringlabels?expand=1#diff-71b2f22455f49b77b1e67f02bb1ff142151a8497c0125345f63b48dca946bb0cL224) but this seems to be ok as the labels should always be sorted when being contracted.

See: https://github.com/prometheus/prometheus/blob/24d7e5bd49dd26a45c2123a897c6cac5ac7c47f0/model/labels/labels.go#L365

cc @fpetkovski