@alanprot 2023-04-28
Waiting https://github.com/cortexproject/promqlsmith/pull/32