@alanprot 2023-05-03
I will run a bench to see the difference.