@alanprot 2023-05-03
old = main
new=stringlabels branch

```
[ec2-user@ip-172-31-30-12 benchmarks]$ ~/go/bin/benchstat old.out new.out
goos: linux
goarch: amd64
pkg: github.com/thanos-community/promql-engine/engine
cpu: Intel(R) Xeon(R) Platinum 8259CL CPU @ 2.50GHz
                                                         │   old.out    │               new.out               │
                                                         │    sec/op    │    sec/op     vs base               │
RangeQuery/vector_selector-32                              38.90m ± ∞ ¹   38.84m ± ∞ ¹        ~ (p=0.690 n=5)
RangeQuery/sum-32                                          13.90m ± ∞ ¹   13.87m ± ∞ ¹        ~ (p=1.000 n=5)
RangeQuery/sum_by_pod-32                                   22.17m ± ∞ ¹   22.17m ± ∞ ¹        ~ (p=0.690 n=5)
RangeQuery/topk-32                                         18.45m ± ∞ ¹   18.44m ± ∞ ¹        ~ (p=1.000 n=5)
RangeQuery/bottomk-32                                      18.72m ± ∞ ¹   18.50m ± ∞ ¹        ~ (p=0.095 n=5)
RangeQuery/rate-32                                         34.52m ± ∞ ¹   34.70m ± ∞ ¹        ~ (p=0.690 n=5)
RangeQuery/rate_with_large_range_selection-32              50.46m ± ∞ ¹   49.89m ± ∞ ¹        ~ (p=0.095 n=5)
RangeQuery/rate_with_large_number_of_series,_1m_range-32    1.422 ± ∞ ¹    1.438 ± ∞ ¹        ~ (p=0.310 n=5)
RangeQuery/rate_with_large_number_of_series,_5m_range-32    1.420 ± ∞ ¹    1.430 ± ∞ ¹   +0.71% (p=0.008 n=5)
RangeQuery/sum_rate-32                                     23.06m ± ∞ ¹   22.95m ± ∞ ¹        ~ (p=0.095 n=5)
RangeQuery/sum_by_rate-32                                  28.07m ± ∞ ¹   28.04m ± ∞ ¹        ~ (p=0.690 n=5)
RangeQuery/quantile_with_variable_parameter-32             35.25m ± ∞ ¹   35.41m ± ∞ ¹   +0.46% (p=0.008 n=5)
RangeQuery/binary_operation_with_one_to_one-32             28.76m ± ∞ ¹   29.23m ± ∞ ¹   +1.61% (p=0.008 n=5)
RangeQuery/binary_operation_with_many_to_one-32            63.25m ± ∞ ¹   65.25m ± ∞ ¹   +3.16% (p=0.008 n=5)
RangeQuery/binary_operation_with_vector_and_scalar-32      52.74m ± ∞ ¹   53.37m ± ∞ ¹   +1.21% (p=0.008 n=5)
RangeQuery/unary_negation-32                               35.84m ± ∞ ¹   36.23m ± ∞ ¹   +1.09% (p=0.008 n=5)
RangeQuery/vector_and_scalar_comparison-32                 53.56m ± ∞ ¹   53.90m ± ∞ ¹   +0.64% (p=0.008 n=5)
RangeQuery/positive_offset_vector-32                       33.39m ± ∞ ¹   33.75m ± ∞ ¹   +1.09% (p=0.008 n=5)
RangeQuery/at_modifier_-32                                 27.68m ± ∞ ¹   27.98m ± ∞ ¹   +1.07% (p=0.008 n=5)
RangeQuery/at_modifier_with_positive_offset_vector-32      27.38m ± ∞ ¹   27.82m ± ∞ ¹   +1.59% (p=0.008 n=5)
RangeQuery/clamp-32                                        65.79m ± ∞ ¹   67.21m ± ∞ ¹   +2.15% (p=0.008 n=5)
RangeQuery/clamp_min-32                                    61.75m ± ∞ ¹   62.99m ± ∞ ¹   +2.01% (p=0.008 n=5)
RangeQuery/complex_func_query-32                           82.93m ± ∞ ¹   85.02m ± ∞ ¹   +2.52% (p=0.008 n=5)
RangeQuery/func_within_func_query-32                       65.55m ± ∞ ¹   67.36m ± ∞ ¹   +2.77% (p=0.008 n=5)
RangeQuery/aggr_within_func_query-32                       65.65m ± ∞ ¹   67.32m ± ∞ ¹   +2.53% (p=0.008 n=5)
RangeQuery/histogram_quantile-32                           155.6m ± ∞ ¹   159.5m ± ∞ ¹   +2.50% (p=0.008 n=5)
RangeQuery/sort-32                                         35.11m ± ∞ ¹   35.92m ± ∞ ¹   +2.31% (p=0.008 n=5)
RangeQuery/sort_desc-32                                    35.10m ± ∞ ¹   35.90m ± ∞ ¹   +2.29% (p=0.008 n=5)
NativeHistograms/selector-32                               170.7m ± ∞ ¹   169.1m ± ∞ ¹        ~ (p=0.841 n=5)
NativeHistograms/sum-32                                    197.9m ± ∞ ¹   200.4m ± ∞ ¹   +1.25% (p=0.032 n=5)
NativeHistograms/rate-32                                   334.8m ± ∞ ¹   297.4m ± ∞ ¹  -11.16% (p=0.008 n=5)
NativeHistograms/sum_rate-32                               348.0m ± ∞ ¹   311.3m ± ∞ ¹  -10.55% (p=0.008 n=5)
NativeHistograms/histogram_sum-32                          472.7m ± ∞ ¹   474.4m ± ∞ ¹        ~ (p=0.421 n=5)
NativeHistograms/histogram_count-32                        471.1m ± ∞ ¹   470.6m ± ∞ ¹        ~ (p=0.690 n=5)
NativeHistograms/histogram_quantile-32                     198.8m ± ∞ ¹   200.2m ± ∞ ¹        ~ (p=0.151 n=5)
geomean                                                    71.38m         71.54m         +0.23%
¹ need >= 6 samples for confidence interval at level 0.95

                                                         │    old.out     │                new.out                │
                                                         │      B/op      │      B/op       vs base               │
RangeQuery/vector_selector-32                               30.17Mi ± ∞ ¹    28.56Mi ± ∞ ¹        ~ (p=1.000 n=5)
RangeQuery/sum-32                                           10.45Mi ± ∞ ¹    10.44Mi ± ∞ ¹        ~ (p=0.151 n=5)
RangeQuery/sum_by_pod-32                                    19.95Mi ± ∞ ¹    19.94Mi ± ∞ ¹        ~ (p=0.222 n=5)
RangeQuery/topk-32                                          12.16Mi ± ∞ ¹    12.11Mi ± ∞ ¹        ~ (p=0.056 n=5)
RangeQuery/bottomk-32                                       12.16Mi ± ∞ ¹    12.14Mi ± ∞ ¹        ~ (p=0.095 n=5)
RangeQuery/rate-32                                          30.69Mi ± ∞ ¹    30.92Mi ± ∞ ¹   +0.74% (p=0.008 n=5)
RangeQuery/rate_with_large_range_selection-32               98.41Mi ± ∞ ¹    98.66Mi ± ∞ ¹   +0.26% (p=0.008 n=5)
RangeQuery/rate_with_large_number_of_series,_1m_range-32   1018.9Mi ± ∞ ¹   1023.8Mi ± ∞ ¹        ~ (p=0.095 n=5)
RangeQuery/rate_with_large_number_of_series,_5m_range-32    1.104Gi ± ∞ ¹    1.108Gi ± ∞ ¹        ~ (p=0.310 n=5)
RangeQuery/sum_rate-32                                      12.36Mi ± ∞ ¹    12.52Mi ± ∞ ¹   +1.28% (p=0.008 n=5)
RangeQuery/sum_by_rate-32                                   21.85Mi ± ∞ ¹    22.05Mi ± ∞ ¹   +0.94% (p=0.008 n=5)
RangeQuery/quantile_with_variable_parameter-32              36.79Mi ± ∞ ¹    36.74Mi ± ∞ ¹   -0.12% (p=0.032 n=5)
RangeQuery/binary_operation_with_one_to_one-32              20.00Mi ± ∞ ¹    20.21Mi ± ∞ ¹   +1.06% (p=0.008 n=5)
RangeQuery/binary_operation_with_many_to_one-32             46.75Mi ± ∞ ¹    47.45Mi ± ∞ ¹   +1.50% (p=0.008 n=5)
RangeQuery/binary_operation_with_vector_and_scalar-32       32.00Mi ± ∞ ¹    32.41Mi ± ∞ ¹   +1.29% (p=0.032 n=5)
RangeQuery/unary_negation-32                                29.33Mi ± ∞ ¹    29.31Mi ± ∞ ¹        ~ (p=0.151 n=5)
RangeQuery/vector_and_scalar_comparison-32                  32.02Mi ± ∞ ¹    31.96Mi ± ∞ ¹        ~ (p=1.000 n=5)
RangeQuery/positive_offset_vector-32                        27.48Mi ± ∞ ¹    27.45Mi ± ∞ ¹        ~ (p=0.310 n=5)
RangeQuery/at_modifier_-32                                  23.04Mi ± ∞ ¹    23.03Mi ± ∞ ¹   -0.02% (p=0.008 n=5)
RangeQuery/at_modifier_with_positive_offset_vector-32       22.85Mi ± ∞ ¹    22.85Mi ± ∞ ¹   -0.02% (p=0.008 n=5)
RangeQuery/clamp-32                                         28.65Mi ± ∞ ¹    28.80Mi ± ∞ ¹   +0.51% (p=0.008 n=5)
RangeQuery/clamp_min-32                                     28.56Mi ± ∞ ¹    28.73Mi ± ∞ ¹   +0.59% (p=0.032 n=5)
RangeQuery/complex_func_query-32                            32.51Mi ± ∞ ¹    33.05Mi ± ∞ ¹   +1.65% (p=0.008 n=5)
RangeQuery/func_within_func_query-32                        30.30Mi ± ∞ ¹    30.68Mi ± ∞ ¹   +1.27% (p=0.008 n=5)
RangeQuery/aggr_within_func_query-32                        30.29Mi ± ∞ ¹    30.74Mi ± ∞ ¹   +1.47% (p=0.008 n=5)
RangeQuery/histogram_quantile-32                            48.26Mi ± ∞ ¹    49.27Mi ± ∞ ¹   +2.09% (p=0.008 n=5)
RangeQuery/sort-32                                          28.35Mi ± ∞ ¹    28.30Mi ± ∞ ¹        ~ (p=0.095 n=5)
RangeQuery/sort_desc-32                                     28.34Mi ± ∞ ¹    28.32Mi ± ∞ ¹        ~ (p=0.151 n=5)
NativeHistograms/selector-32                                718.4Mi ± ∞ ¹    718.3Mi ± ∞ ¹        ~ (p=0.548 n=5)
NativeHistograms/sum-32                                     879.0Mi ± ∞ ¹    878.9Mi ± ∞ ¹        ~ (p=0.690 n=5)
NativeHistograms/rate-32                                    1.304Gi ± ∞ ¹    1.102Gi ± ∞ ¹  -15.52% (p=0.008 n=5)
NativeHistograms/sum_rate-32                                1.460Gi ± ∞ ¹    1.258Gi ± ∞ ¹  -13.86% (p=0.008 n=5)
NativeHistograms/histogram_sum-32                           720.0Mi ± ∞ ¹    720.1Mi ± ∞ ¹   +0.02% (p=0.008 n=5)
NativeHistograms/histogram_count-32                         720.0Mi ± ∞ ¹    720.2Mi ± ∞ ¹   +0.03% (p=0.032 n=5)
NativeHistograms/histogram_quantile-32                      879.4Mi ± ∞ ¹    879.3Mi ± ∞ ¹        ~ (p=0.421 n=5)
geomean                                                     66.73Mi          66.29Mi         -0.66%
¹ need >= 6 samples for confidence interval at level 0.95

                                                         │   old.out    │               new.out               │
                                                         │  allocs/op   │  allocs/op    vs base               │
RangeQuery/vector_selector-32                              125.1k ± ∞ ¹   101.7k ± ∞ ¹        ~ (p=1.000 n=5)
RangeQuery/sum-32                                          83.15k ± ∞ ¹   82.99k ± ∞ ¹   -0.19% (p=0.008 n=5)
RangeQuery/sum_by_pod-32                                   180.5k ± ∞ ¹   180.3k ± ∞ ¹   -0.09% (p=0.008 n=5)
RangeQuery/topk-32                                         102.7k ± ∞ ¹   101.9k ± ∞ ¹   -0.72% (p=0.008 n=5)
RangeQuery/bottomk-32                                      103.9k ± ∞ ¹   103.8k ± ∞ ¹   -0.10% (p=0.008 n=5)
RangeQuery/rate-32                                         120.5k ± ∞ ¹   120.7k ± ∞ ¹   +0.20% (p=0.008 n=5)
RangeQuery/rate_with_large_range_selection-32              113.3k ± ∞ ¹   113.7k ± ∞ ¹   +0.38% (p=0.008 n=5)
RangeQuery/rate_with_large_number_of_series,_1m_range-32   3.703M ± ∞ ¹   3.704M ± ∞ ¹   +0.02% (p=0.008 n=5)
RangeQuery/rate_with_large_number_of_series,_5m_range-32   3.903M ± ∞ ¹   3.904M ± ∞ ¹   +0.02% (p=0.032 n=5)
RangeQuery/sum_rate-32                                     106.0k ± ∞ ¹   106.2k ± ∞ ¹   +0.16% (p=0.040 n=5)
RangeQuery/sum_by_rate-32                                  203.5k ± ∞ ¹   203.8k ± ∞ ¹   +0.11% (p=0.008 n=5)
RangeQuery/quantile_with_variable_parameter-32             535.3k ± ∞ ¹   535.0k ± ∞ ¹   -0.05% (p=0.008 n=5)
RangeQuery/binary_operation_with_one_to_one-32             90.56k ± ∞ ¹   92.52k ± ∞ ¹   +2.16% (p=0.008 n=5)
RangeQuery/binary_operation_with_many_to_one-32            158.6k ± ∞ ¹   164.6k ± ∞ ¹   +3.78% (p=0.008 n=5)
RangeQuery/binary_operation_with_vector_and_scalar-32      101.3k ± ∞ ¹   104.4k ± ∞ ¹   +3.00% (p=0.008 n=5)
RangeQuery/unary_negation-32                               108.9k ± ∞ ¹   108.7k ± ∞ ¹   -0.16% (p=0.008 n=5)
RangeQuery/vector_and_scalar_comparison-32                 101.4k ± ∞ ¹   101.2k ± ∞ ¹   -0.19% (p=0.016 n=5)
RangeQuery/positive_offset_vector-32                       85.36k ± ∞ ¹   85.21k ± ∞ ¹   -0.18% (p=0.032 n=5)
RangeQuery/at_modifier_-32                                 63.66k ± ∞ ¹   63.60k ± ∞ ¹   -0.09% (p=0.008 n=5)
RangeQuery/at_modifier_with_positive_offset_vector-32      57.66k ± ∞ ¹   57.60k ± ∞ ¹   -0.10% (p=0.008 n=5)
RangeQuery/clamp-32                                        100.9k ± ∞ ¹   103.7k ± ∞ ¹   +2.77% (p=0.008 n=5)
RangeQuery/clamp_min-32                                    100.4k ± ∞ ¹   103.2k ± ∞ ¹   +2.83% (p=0.008 n=5)
RangeQuery/complex_func_query-32                           105.5k ± ∞ ¹   111.5k ± ∞ ¹   +5.73% (p=0.008 n=5)
RangeQuery/func_within_func_query-32                       121.0k ± ∞ ¹   124.1k ± ∞ ¹   +2.62% (p=0.008 n=5)
RangeQuery/aggr_within_func_query-32                       120.9k ± ∞ ¹   124.3k ± ∞ ¹   +2.74% (p=0.008 n=5)
RangeQuery/histogram_quantile-32                           541.1k ± ∞ ¹   563.5k ± ∞ ¹   +4.12% (p=0.008 n=5)
RangeQuery/sort-32                                         99.90k ± ∞ ¹   99.67k ± ∞ ¹   -0.23% (p=0.008 n=5)
RangeQuery/sort_desc-32                                    99.87k ± ∞ ¹   99.69k ± ∞ ¹   -0.19% (p=0.008 n=5)
NativeHistograms/selector-32                               9.593M ± ∞ ¹   9.592M ± ∞ ¹        ~ (p=0.421 n=5)
NativeHistograms/sum-32                                    11.03M ± ∞ ¹   11.03M ± ∞ ¹        ~ (p=0.968 n=5)
NativeHistograms/rate-32                                   22.72M ± ∞ ¹   17.72M ± ∞ ¹  -22.01% (p=0.008 n=5)
NativeHistograms/sum_rate-32                               24.15M ± ∞ ¹   19.15M ± ∞ ¹  -20.71% (p=0.008 n=5)
NativeHistograms/histogram_sum-32                          9.593M ± ∞ ¹   9.596M ± ∞ ¹   +0.03% (p=0.008 n=5)
NativeHistograms/histogram_count-32                        9.593M ± ∞ ¹   9.596M ± ∞ ¹   +0.03% (p=0.008 n=5)
NativeHistograms/histogram_quantile-32                     11.03M ± ∞ ¹   11.03M ± ∞ ¹        ~ (p=0.222 n=5)
geomean                                                    374.2k         369.8k         -1.16%
¹ need >= 6 samples for confidence interval at level 0.95
```